Prevent an Application's child resources from being deleted, when the parent Application is deleted
By default, when an Application resource is deleted by the
ApplicationSet controller, all of the child resources of the Application
will be deleted as well (such as, all of the Application's Deployments,
Services, etc).

To prevent an Application's child resources from being deleted when the
parent Application is deleted, add the preserveResourcesOnDeletion: true
field to the syncPolicy of the ApplicationSet.

See also:
https://argocd-applicationset.readthedocs.io/en/stable/Controlling-Resource-Modification/